Added RecyclerView Manager, background. changed main icon

+++ app/src/main/java/com/azimut4946777/shoptask/MainTaskFragment.kt
@@ -1,18 +1,30 @@
 package com.azimut4946777.shoptask
 
+import com.azimut4946777.shoptask.*
+
+
+
 import android.os.Bundle
+import android.util.Log
 import android.view.*
-import androidx.appcompat.app.AppCompatActivity
+import androidx.core.graphics.alpha
+import androidx.core.view.isEmpty
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import androidx.navigation.findNavController
 import androidx.navigation.fragment.findNavController
+import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.StaggeredGridLayoutManager
+import com.azimut4946777.shoptask.data.Task
+import com.azimut4946777.shoptask.data.TaskRoomDatabase
 import com.azimut4946777.shoptask.databinding.FragmentMainTaskBinding
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
+
 
 
 class MainTaskFragment : Fragment() {
+
 
     private val viewModel: ShopTaskViewModel by activityViewModels {
         ShopTaskViewModelFactory(
@@ -21,6 +33,8 @@ class MainTaskFragment : Fragment() {
     }
 
 
+    // For Layout Manager
+    private var isLinearLayoutManager = true
     private var _binding: FragmentMainTaskBinding? = null
     private val binding get() = _binding!!
 
@@ -36,11 +50,6 @@ class MainTaskFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         _binding = FragmentMainTaskBinding.inflate(layoutInflater, container, false)
-
-
-
-
-
         return binding.root
     }
 
@@ -50,20 +59,28 @@ class MainTaskFragment : Fragment() {
             val action = MainTaskFragmentDirections.actionMainTaskFragmentToEditTaskFragment(it.id)
             this.findNavController().navigate(action)
         }
-        binding.taskRecyclerview.adapter = adapter
+
+        val recyclerView = binding.taskRecyclerview
+
+        recyclerView.adapter = adapter
+
+        recyclerView.layoutManager = LinearLayoutManager(context)
+
         viewModel.allTasks.observe(this.viewLifecycleOwner) { items ->
             items.let {
                 adapter.submitList(it)
             }
         }
-        binding.taskRecyclerview.layoutManager = LinearLayoutManager(this.context)
+            //For CheckBox
+
+
         //FAB bottom action
         binding.fab.setOnClickListener {
             val action = MainTaskFragmentDirections.actionMainTaskFragmentToAddTaskFragment()
             findNavController().navigate(action)
         }
         //Hide FAB
-        binding.taskRecyclerview.addOnScrollListener(object : RecyclerView.OnScrollListener() {
+        recyclerView.addOnScrollListener(object : RecyclerView.OnScrollListener() {
             override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
                 super.onScrolled(recyclerView, dx, dy)
 
@@ -92,23 +109,63 @@ class MainTaskFragment : Fragment() {
     //For AppBar
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
         inflater.inflate(R.menu.top_menu, menu)
+
     }
 
     //For AppBar. Button Action
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         return when (item.itemId) {
             R.id.delete_all -> {
-                viewModel.deleteAllTask()
+                if (binding.taskRecyclerview.isEmpty()) {
+                }else {
+                    showDeleteAllDialog()
+                }
+
                 true
             }
             R.id.grid -> {
-
+                // For Layout Manager
+                isLinearLayoutManager = !isLinearLayoutManager
+                chooseLayout()
+                setIcon(item)
                 true
             }
             else -> super.onOptionsItemSelected(item)
         }
-
     }
+
+    // For Layout Manager. Change RecyclerView LayoutManager
+    private fun chooseLayout() {
+        if (isLinearLayoutManager) {
+            binding.taskRecyclerview.layoutManager = LinearLayoutManager(context)
+        } else {
+            //binding.taskRecyclerview.layoutManager = GridLayoutManager(context, 2)
+            binding.taskRecyclerview.layoutManager = StaggeredGridLayoutManager(2,LinearLayoutManager.VERTICAL)
+        }
+    }
+
+    // For Layout Manager. Change Icon
+    private fun setIcon(menuItem: MenuItem) {
+        if (isLinearLayoutManager) {
+            menuItem.setIcon(R.drawable.ic_linear_layout)
+        } else {
+            menuItem.setIcon(R.drawable.ic_grid_layout)
+        }
+    }
+
+
+    private fun showDeleteAllDialog() {
+        MaterialAlertDialogBuilder(requireContext())
+            .setTitle(getString(R.string.alert_title))
+            .setMessage(getString(R.string.delete_question))
+            .setCancelable(false)
+            .setNegativeButton(getString(R.string.no)){_,_->}
+            .setPositiveButton(getString(R.string.yes)){_,_->
+                viewModel.deleteAllTask()
+            }
+            .show()
+    }
+
 
     override fun onDestroyView() {
         super.onDestroyView()
